requirements.txt
# Core scientific stack
# ipython
# jupyter
openpyxl
pandas
geopandas
rioxarray
dask[complete]
cartopy
colormaps
matplotlib
scipy
# h5netcdf
# netCDF4
pymannkendall
tqdm
zarr
# ipyleaflet
wetterdienst
pint-pandas
cdsapi
isimip-client

# Workflow + config
hydra-core
aiohttp
requests
gcsfs
dask-jobqueue
seaborn

# Earth Engine / remote sensing
earthengine-api
geemap
xarray
xesmf
intake
intake-esm

# Climate/statistics packages
xsdba 
xclim
pyarrow
fastparquet

# ML stack
mlflow
scikit-learn
xgboost  # GPU support may not work in Colab, CPU version installs fine
optuna

# Dev tools
gitpython
beautifulsoup4
google-auth
google-api-python-client
# ipdb

# imputegap (vendorized)

toml
tensorflow
shap
pycatch22
scikit-optimize
pyswarms
pytest
xlsxwriter
types-toml
types-setuptools
wheel
sktime
statsmodels
prophet
plotly
darts
torchvision
torchaudio
ujson
ipdb
torchcde
tables
MissForest
lightgbm
tsfel
tsfresh
einops
ray
fire
fastdtw
torch-geometric
pytorch-lightning
torchmetrics
lightning
reformer-pytorch
transformers
timm
pyqt5
